# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HIMS Hospital Information Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List / search patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Patient list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListResponse"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient demographics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientBody"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"

  # ── APPOINTMENTS / OPD QUEUE ──────────────────────────────────────────────
...
  /lab-orders:
    get:
      operationId: listLabOrders
      tags: [labOrders]
      summary: List lab orders
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lab order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrderListResponse"
    post:
      operationId: createLabOrder
      tags: [labOrders]
      summary: Order lab tests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabOrderBody"
      responses:
        "201":
          description: Lab order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
  /lab-orders/{id}:
    get:
      operationId: getLabOrder
      tags: [labOrders]
      summary: Get lab order details and results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lab order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
    patch:
      operationId: updateLabOrder
      tags: [labOrders]
      summary: Enter lab results or update status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabOrderBody"
      responses:
        "200":
          description: Updated lab order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"

  # ── PRESCRIPTIONS ────────────────────────────────────────────────────────
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: consultationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Prescription list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionListResponse"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Create a prescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionBody"
      responses:
        "201":
          description: Prescription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"

  # ── MEDICATIONS (Pharmacy Drug Catalog) ──────────────────────────────────
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [billing]
      summary: Generate invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [billing]
      summary: Get invoice details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [billing]
      summary: Update invoice status (pay, void)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ── EMERGENCY ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        mrNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        bloodGroup:
          type: ["string", "null"]
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        chronicConditions:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - mrNumber
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phone
        - createdAt
        - updatedAt
    CreatePatientBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        bloodGroup:
          type: ["string", "null"]
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        chronicConditions:
          type: ["string", "null"]
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phone
    UpdatePatientBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        bloodGroup:
          type: ["string", "null"]
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        chronicConditions:
          type: ["string", "null"]
    PatientListResponse:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
      required:
        - patients
        - total

    # ── APPOINTMENT ─────────────────────────────────────────────────────────
...
    LabOrder:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        doctorId:
          type: integer
        doctorName:
          type: string
        consultationId:
          type: ["integer", "null"]
        testName:
          type: string
        testCode:
          type: ["string", "null"]
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        result:
          type: ["string", "null"]
        normalRange:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        resultNotes:
          type: ["string", "null"]
        orderedAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - patientId
        - patientName
        - doctorId
        - doctorName
        - testName
        - category
        - priority
        - status
        - orderedAt
    CreateLabOrderBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        consultationId:
          type: ["integer", "null"]
        testName:
          type: string
        testCode:
          type: ["string", "null"]
        category:
          type: string
        priority:
          type: string
      required:
        - patientId
        - doctorId
        - testName
        - category
        - priority
    UpdateLabOrderBody:
      type: object
      properties:
        status:
          type: string
        result:
          type: ["string", "null"]
        normalRange:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        resultNotes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
    LabOrderListResponse:
      type: object
      properties:
        labOrders:
          type: array
          items:
            $ref: "#/components/schemas/LabOrder"
        total:
          type: integer
      required:
        - labOrders
        - total

    # ── PRESCRIPTION ─────────────────────────────────────────────────────────
    Prescription:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        doctorId:
          type: integer
        doctorName:
          type: string
        consultationId:
          type: ["integer", "null"]
        drugName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        route:
          type: string
        instructions:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - patientId
        - patientName
        - doctorId
        - doctorName
        - drugName
        - dosage
        - frequency
        - duration
        - route
        - status
        - createdAt
    CreatePrescriptionBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        consultationId:
          type: ["integer", "null"]
        drugName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        route:
          type: string
        instructions:
          type: ["string", "null"]
      required:
        - patientId
        - doctorId
        - drugName
        - dosage
        - frequency
        - duration
        - route
    PrescriptionListResponse:
      type: object
      properties:
        prescriptions:
          type: array
          items:
            $ref: "#/components/schemas/Prescription"
        total:
          type: integer
      required:
        - prescriptions
        - total

    # ── MEDICATION ───────────────────────────────────────────────────────────
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        patientId:
          type: integer
        patientName:
          type: string
        admissionId:
          type: ["integer", "null"]
        consultationId:
          type: ["integer", "null"]
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        status:
          type: string
        paymentMethod:
          type: ["string", "null"]
        items:
          type: string
        createdAt:
          type: string
        paidAt:
          type: ["string", "null"]
      required:
        - id
        - invoiceNumber
        - patientId
        - patientName
        - subtotal
        - discount
        - tax
        - total
        - status
        - items
        - createdAt
    CreateInvoiceBody:
      type: object
      properties:
        patientId:
          type: integer
        admissionId:
          type: ["integer", "null"]
        consultationId:
          type: ["integer", "null"]
        items:
          type: string
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        paymentMethod:
          type: ["string", "null"]
      required:
        - patientId
        - items
        - subtotal
        - discount
        - tax
        - total
    UpdateInvoiceBody:
      type: object
      properties:
        status:
          type: string
        paymentMethod:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
    InvoiceListResponse:
      type: object
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
      required:
        - invoices
        - total

    # ── EMERGENCY ────────────────────────────────────────────────────────────
...